REF: Move `AnalysisLogFormatterSwitcher`
